requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
python-docx>=0.8.11
PyPDF2>=3.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-decouple>=3.8
djangorestframework>=3.14.0 